refactor: renderResultSectionContent 파일을 페이지 단위로 분리
